+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.12.0
+
+- Bump `rarena-allocator`'s version
+- Change value callback from `impl FnOnce + Copy` to `impl Fn`
+- Give users back the key offset in the ARENA and the key in value callback
+
 ## 0.11.0
 
 - Refactor and extract lock-free ARENA allocator implementation to [`rarena-allocator`](https://github.com/al8n/rarena) crate.
@@ -78,5 +84,4 @@
 - Add `flush` API
 
 ## UNRELEASED
-